Added edit message, fix bugg for attachments

+++ cogs/logging.py
@@ -4,13 +4,12 @@
 import os
 
 import discord
-from discord import guild, Message
+from discord import Message
 from discord.ext import commands
-from discord.ext.commands import bot
 
+from cogs.embeds.UpdatedMessageEmbed import UpdatedMessageEmbed
 from cogs.embeds.deleted_message_embed import DeleteMessageEmbed
 from cogs.embeds.join_embed import JoinEmbed
-from cogs.embeds.message_embed import MessageEmbed
 from cogs.embeds.sent_message_embed import SentMessageEmbed
 
 
@@ -39,8 +38,13 @@ class Logging(commands.Cog):
 
     @commands.Cog.listener()
     async def on_message(self, message: Message):
-        # TODO: FIX FILES
-        if not message.author.bot:
+        """"
+        Event listener when a message is deleted.
+        Only works with the internal cache of the bot, so won't trigger if the bot hasn't cached the message.
+        """
+
+        # Check for message.content, in case it is only a attachment
+        if not message.author.bot and message.content:
             message_embed = SentMessageEmbed(message=message)
             message_embed.build_embed()
 
@@ -54,8 +58,9 @@ class Logging(commands.Cog):
         Event listener when a message is deleted.
         Only works with the internal cache of the bot, so won't trigger if the bot hasn't cached the message.
         """
-        # TODO: FIX FILES
-        if not message.author.bot:
+
+        # Check for message.content, in case it is only a attachment
+        if not message.author.bot and message.content:
             message_embed = DeleteMessageEmbed(message=message)
             message_embed.build_embed()
 
@@ -63,6 +68,23 @@ class Logging(commands.Cog):
                 embed=message_embed.embed
             )
 
+    @commands.Cog.listener()
+    async def on_message_edit(self, original_message: Message, updated_message: Message):
+        """"
+        Event listener when a message is updated.
+        Only works with the internal cache of the bot, so won't trigger if the bot hasn't cached the message.
+        """
+
+        # Check for message.content, in case it is only a attachment
+        if not original_message.author.bot:
+            original_message_has_content = True if original_message.content else False
+            message_embed = UpdatedMessageEmbed(original_message=original_message, updated_message=updated_message,
+                                                original_message_has_content=original_message_has_content)
+            message_embed.build_embed()
+
+            await self.channel.send(
+                embed=message_embed.embed
+            )
 
 
 def setup(bot):
